+++ lib/AppFramework/BusinessLayer/BusinessLayer.php
@@ -209,7 +209,7 @@ abstract class BusinessLayer {
 	 * 				(see https://ampache.org/api/api-advanced-search#available-search-rules, alias names not supported here),
 	 * 				'operator' is one of 
 	 * 				['contain', 'notcontain', 'start', 'end', 'is', 'isnot', 'sounds', 'notsounds', 'regexp', 'notregexp',
-	 * 				 '>=', '<=', '=', '!=', '>', '<', 'true', 'false', 'equal', 'ne', 'limit'],
+	 * 				 '>=', '<=', '=', '!=', '>', '<', 'before', 'after', 'true', 'false', 'equal', 'ne', 'limit'],
 	 * 				'input' is the right side value of the 'operator' (disregarded for the operators 'true' and 'false')
 	 * @param Random $random When the randomization utility is passed, the result set will be in random order (still supporting proper paging).
 	 * 						 In this case, the argument $sortBy is ignored.